Split the import resolver into its own module
